feat: merge readlog into log command
